feat: fixed the security issue for dynamic update
